// src/types/retailer.ts
export interface RetailerOnboardingData {
  brandName: string;
  brandLogo: string;
  brandWebsite: string;
  offerTitle: string;
  offerDescription: string;
  discountValue: string;
  discountCode: string;
  isUniqueCode: boolean;
  uniqueCodes: string[];
  expiryDate: string;
  terms: string[];
  useAffiliateLinks: boolean;
  affiliateNetwork: string;
  quote: string;
  joinEmployeeDiscountBeta: boolean;
}